lbipam: split out LBIPAM config into pkg/lbipamconfig
Split configuration out of LBIPAM into its own cell so that other
packages can pull in LBIPAM configuration.

The rationale to position this outside the operator is to avoid
shared packages	building dependencies into the operator.
