+++ pkg/metrics/features/operator/cell.go
@@ -13,8 +13,8 @@ import (
 	operatorOption "github.com/cilium/cilium/operator/option"
 	ciliumenvoyconfig2 "github.com/cilium/cilium/operator/pkg/ciliumenvoyconfig"
 	"github.com/cilium/cilium/operator/pkg/ingress"
-	"github.com/cilium/cilium/operator/pkg/lbipam"
 	k8sversion "github.com/cilium/cilium/pkg/k8s/version"
+	"github.com/cilium/cilium/pkg/lbipamconfig"
 	"github.com/cilium/cilium/pkg/metrics"
 	"github.com/cilium/cilium/pkg/nodeipamconfig"
 )
@@ -57,7 +57,7 @@ type featuresParams struct {
 	OperatorConfig *operatorOption.OperatorConfig
 
 	IngressController ingress.IngressConfig
-	LBIPAM            lbipam.Config
+	LBIPAM            lbipamconfig.Config
 	LBConfig          ciliumenvoyconfig2.LoadBalancerConfig
 	NodeIPAM          nodeipamconfig.NodeIPAMConfig
 }
